Reduce number of Spectral Pybindings files
