+++ src/NumericalAlgorithms/Spectral/Python/Bindings.cpp
@@ -14,11 +14,7 @@ PYBIND11_MODULE(_Pybindings, m) {  // NOLINT
   enable_segfault_handler();
   py::module_::import("spectre.DataStructures");
   py::module_::import("spectre.DataStructures.Tensor");
-  Spectral::py_bindings::bind_basis(m);
-  Spectral::py_bindings::bind_quadrature(m);
   Spectral::py_bindings::bind_logical_coordinates(m);
-  Spectral::py_bindings::bind_nodal_to_modal_matrix(m);
-  Spectral::py_bindings::bind_modal_to_nodal_matrix(m);
-  Spectral::py_bindings::bind_collocation_points(m);
+  Spectral::py_bindings::bind_spectral(m);
   py_bindings::bind_mesh(m);
 }
