+++ src/NumericalAlgorithms/Spectral/Python/Spectral.hpp
@@ -7,9 +7,5 @@
 
 namespace Spectral::py_bindings {
 // NOLINTNEXTLINE(google-runtime-references)
-void bind_basis(pybind11::module& m);
-void bind_quadrature(pybind11::module& m);
-void bind_modal_to_nodal_matrix(pybind11::module& m);
-void bind_nodal_to_modal_matrix(pybind11::module& m);
-void bind_collocation_points(pybind11::module& m);
+void bind_spectral(pybind11::module& m);
 }  // namespace Spectral::py_bindings
