+++ tests/Unit/NumericalAlgorithms/Spectral/Python/Test_Spectral.py
@@ -7,10 +7,17 @@ import numpy as np
 from numpy.polynomial import chebyshev, legendre
 
 from spectre.DataStructures import DataVector
-from spectre.Spectral import Basis, Mesh1D, Quadrature, collocation_points
+from spectre.Spectral import (
+    Basis,
+    Mesh1D,
+    Quadrature,
+    collocation_points,
+    modal_to_nodal_matrix,
+    nodal_to_modal_matrix,
+)
 
 
-class TestCollocationPoints(unittest.TestCase):
+class TestSpectral(unittest.TestCase):
     def test_collocation_points_legendre_gauss(self):
         for num_coll_points in range(2, 13):
             coefs = np.zeros(num_coll_points + 1)
@@ -62,6 +69,40 @@ class TestCollocationPoints(unittest.TestCase):
             np.testing.assert_allclose(
                 coll_points_numpy, coll_points_spectre, 1e-12, 1e-12
             )
+
+    def test_modal_to_nodal_matrix_legendre(self):
+        for quadrature in [Quadrature.Gauss, Quadrature.GaussLobatto]:
+            for num_coll_points in range(2, 13):
+                mesh = Mesh1D(num_coll_points, Basis.Legendre, quadrature)
+                coll_points = collocation_points(mesh)
+                mtn_numpy = legendre.legvander(coll_points, num_coll_points - 1)
+                mtn_spectre = modal_to_nodal_matrix(mesh)
+                np.testing.assert_allclose(mtn_spectre, mtn_numpy, 1e-12, 1e-12)
+                ntm_spectre = nodal_to_modal_matrix(mesh)
+                np.testing.assert_allclose(
+                    np.matmul(mtn_spectre, ntm_spectre),
+                    np.identity(num_coll_points),
+                    1e-12,
+                    1e-12,
+                )
+
+    def test_modal_to_nodal_matrix_chebyshev(self):
+        for quadrature in [Quadrature.Gauss, Quadrature.GaussLobatto]:
+            for num_coll_points in range(2, 13):
+                mesh = Mesh1D(num_coll_points, Basis.Chebyshev, quadrature)
+                coll_points = collocation_points(mesh)
+                mtn_numpy = chebyshev.chebvander(
+                    coll_points, num_coll_points - 1
+                )
+                mtn_spectre = modal_to_nodal_matrix(mesh)
+                np.testing.assert_allclose(mtn_spectre, mtn_numpy, 1e-12, 1e-12)
+                ntm_spectre = nodal_to_modal_matrix(mesh)
+                np.testing.assert_allclose(
+                    np.matmul(mtn_spectre, ntm_spectre),
+                    np.identity(num_coll_points),
+                    1e-12,
+                    1e-12,
+                )
 
 
 if __name__ == "__main__":
